배포를 위한 병합

+++ src/main/java/com/nhnacademy/hexashoppingmallservice/controller/member/MemberCouponController.java
@@ -2,6 +2,8 @@ package com.nhnacademy.hexashoppingmallservice.controller.member;
 
 import com.nhnacademy.hexashoppingmallservice.entity.member.MemberCoupon;
 import com.nhnacademy.hexashoppingmallservice.service.member.MemberCouponService;
+import com.nhnacademy.hexashoppingmallservice.util.JwtUtils;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,14 +17,17 @@ import java.util.List;
 public class MemberCouponController {
     private final MemberCouponService memberCouponService;
 
+    private final JwtUtils jwtUtils;
+
     /**
      * 특정 회원의 쿠폰 목록을 조회하는 엔드포인트
      *
      * @param memberId 쿠폰을 조회할 회원의 ID
      * @return 회원의 쿠폰 목록
      */
-    @GetMapping("/auth/members/{memberId}/coupons")
-    public ResponseEntity<List<MemberCoupon>> getMemberCoupons(@PathVariable String memberId) {
+    @GetMapping("/members/{memberId}/coupons")
+    public ResponseEntity<List<MemberCoupon>> getMemberCoupons(@PathVariable String memberId, HttpServletRequest request) {
+        jwtUtils.ensureUserAccess(request, memberId);
         List<MemberCoupon> memberCoupons = memberCouponService.getMemberCoupon(memberId);
         return ResponseEntity.ok(memberCoupons);
     }
@@ -34,16 +39,18 @@ public class MemberCouponController {
      * @param couponId 생성할 쿠폰의 ID
      * @return 생성된 MemberCoupon
      */
-    @PostMapping("/auth/members/{memberId}/coupons/{couponId}")
+    @PostMapping("/members/{memberId}/coupons/{couponId}")
     public ResponseEntity<MemberCoupon> createMemberCoupon(
             @PathVariable String memberId,
-            @PathVariable Long couponId) {
+            @PathVariable Long couponId, HttpServletRequest request) {
+        jwtUtils.ensureAdmin(request);
         MemberCoupon memberCoupon = memberCouponService.createMemberCoupon(couponId, memberId);
         return new ResponseEntity<>(memberCoupon, HttpStatus.CREATED);
     }
 
-    @DeleteMapping("/auth/members/{memberId}/coupons/{couponId}")
-    public ResponseEntity<Void> deleteMemberCoupon(@PathVariable String memberId, @PathVariable Long couponId) {
+    @DeleteMapping("/members/{memberId}/coupons/{couponId}")
+    public ResponseEntity<Void> deleteMemberCoupon(@PathVariable String memberId, @PathVariable Long couponId, HttpServletRequest request) {
+        jwtUtils.ensureAdmin(request);
         memberCouponService.deleteMemberCoupon(couponId, memberId);
         return ResponseEntity.noContent().build();
     }
